# Conflicts:
#	go.mod
#	go.sum
#	portal-ui/bindata_assetfs.go
#	portal-ui/src/screens/Console/Buckets/ViewBucket/ViewBucket.tsx
#	portal-ui/src/screens/Console/Common/FormComponents/SelectWrapper/SelectWrapper.tsx
#	restapi/embedded_spec.go
#	restapi/operations/console_api.go
#	restapi/user_buckets.go
#	swagger.yml

+++ models/error.go
@@ -35,7 +35,7 @@ import (
 type Error struct {
 
 	// code
-	Code int64 `json:"code,omitempty"`
+	Code int32 `json:"code,omitempty"`
 
 	// message
 	// Required: true
